TASK: Re-organize ck editor adapter file structure
