+++ Resources/Private/JavaScript/Guest/Components/Editors/CKEditor/Factory/Selection/index.js
