+++ Resources/Private/JavaScript/Guest/Components/Editors/CKEditor/Factory/Toolbar/Helper/index.js
