+++ Resources/Private/JavaScript/Guest/Components/Editors/CKEditor/Factory/Toolbar/index.js
